requirements/base.txt
celery==5.1.2
Django==3.2.9
django-celery-beat==2.2.1
django-filter==21.1
django-simple-history==3.0.0
djangorestframework==3.12.4
drf-spectacular[sidecar]==0.20.2
factory-boy==3.2.0
psycopg2-binary==2.9.1
pytz==2021.1
redis==3.5.3
requests==2.26.0
